Updating aya-rs/aya-log references to aya-rs/aya
This crate has beed moved from aya-rs/aya-log into the aya repository as
of July-ish this year according to the aya-log git history.  Several of
the examples in the book were not working as expected (myapp-01 for
example was not emitting the format string for ebpf logging macro
invocations) so by updating these Cargo.toml deps to point at the new
crate version we can get the examples back to working out-of-the-box.

This will also help avoid having potential future users depending on the
old/deprecated version of this crate.  I would suggest also adding
additional instructions to the old crate's README.