examples/{tc,xdp}*: Use network-types crate instead of aya-tool bindings
Network-related structures don't change across kernel versions, so using
BTF to generate them is an overkill. Instead of using aya-tool, we
decided to publish a crate called network-types, which provides
definitions of network headers (L2, L3 and L4) which are suitable to use
as compatible with ones provided by the kernel.

Signed-off-by: Michal Rostecki <vadorovsky@gmail.com>