Link and wording fix: lsm.md (#161)
lsm_hooks.h does not appear to actually list the hooks as stated
in the original text, rather it contains definitions for a data structure
providing the list during execution. I also found the wording to be
slightly confusing, as to my understanding, the author of the LSM
module is the one making the policy decisions.